Modified usage of RootAnalysis.version. Added update_root_analysis API call.
The update_root_analysis() call will attempt to update the root if the
version matches, otherwise it returns an error code.

track_root_analysis() also modified to attempt to insert first, update
if it already exists, and return success or failure status.

The version property is updated by the interface API call if the insert
or update is successful.

+++ ace/system/threaded/analysis_tracking.py
@@ -1,6 +1,8 @@
 # vim: ts=4:sw=4:et:cc=120
 
 import json
+import threading
+import uuid
 
 from dataclasses import dataclass, field
 from typing import Union, List, Optional, Any
@@ -21,52 +23,92 @@ class ThreadedAnalysisTrackingInterface(AnalysisTrackingInterface):
 
     root_analysis = {}  # key = RootAnalysis.uuid, value = RootAnalysisTracking
     analysis_details = {}  # key = Analysis.uuid, value = Any
+    sync_lock = threading.RLock()
 
-    def track_root_analysis(self, root: RootAnalysis):
+    def track_root_analysis(self, root: RootAnalysis) -> bool:
         assert isinstance(root, RootAnalysis)
-        self.root_analysis[root.uuid] = RootAnalysisTracking(root=root.to_json(exclude_analysis_details=True))
+        with self.sync_lock:
+            if root.uuid in self.root_analysis:
+                return False
+
+            # assign a version if we don't have one yet
+            if root.version is None:
+                root.version = str(uuid.uuid4())
+
+            self.root_analysis[root.uuid] = RootAnalysisTracking(root=root.to_json(exclude_analysis_details=True))
+            return True
+
+    def update_root_analysis(self, root: RootAnalysis) -> bool:
+        assert isinstance(root, RootAnalysis)
+        with self.sync_lock:
+            # if the root does not already exist then we can't update it
+            existing_root = self.root_analysis.get(root.uuid, None)
+            if not existing_root:
+                return False
+
+            existing_root = RootAnalysis.from_json(existing_root.root)
+            # returning None here indicates that the version of the Root is out of date
+            if existing_root.version != root.version:
+                return False
+
+            # at this point the version changes to something else
+            root.version = str(uuid.uuid4())
+            self.root_analysis[root.uuid] = RootAnalysisTracking(root=root.to_json(exclude_analysis_details=True))
+
+            # and we return the value of the new version
+            return True
 
     def get_root_analysis(self, uuid: str) -> Union[RootAnalysis, None]:
-        try:
-            tracking = self.root_analysis[uuid]
-        except KeyError:
-            return None
+        with self.sync_lock:
+            try:
+                tracking = self.root_analysis[uuid]
+            except KeyError:
+                return None
 
         return RootAnalysis.from_json(tracking.root)
 
     def delete_root_analysis(self, uuid: str) -> bool:
-        root_tracking = self.root_analysis.pop(uuid, None)
+        with self.sync_lock:
+            root_tracking = self.root_analysis.pop(uuid, None)
+
         if not root_tracking:
             return False
 
-        for analysis_uuid in root_tracking.details:
-            self.delete_analysis_details(analysis_uuid)
+        with self.sync_lock:
+            for analysis_uuid in root_tracking.details:
+                self.delete_analysis_details(analysis_uuid)
 
         return True
 
     def root_analysis_exists(self, uuid: str) -> bool:
-        return uuid in self.root_analysis
+        with self.sync_lock:
+            return uuid in self.root_analysis
 
     def get_analysis_details(self, uuid: dict) -> Any:
-        details_json = self.analysis_details.get(uuid)
+        with self.sync_lock:
+            details_json = self.analysis_details.get(uuid)
+
         if details_json is None:
             return None
 
         return json.loads(details_json)
 
     def track_analysis_details(self, root_uuid: str, uuid: str, value):
-        try:
-            self.root_analysis[root_uuid].details.append(uuid)
-        except KeyError:
-            raise UnknownRootAnalysisError(root_uuid)
+        with self.sync_lock:
+            try:
+                self.root_analysis[root_uuid].details.append(uuid)
+            except KeyError:
+                raise UnknownRootAnalysisError(root_uuid)
 
-        self.analysis_details[uuid] = json.dumps(value)
+            self.analysis_details[uuid] = json.dumps(value)
 
     def delete_analysis_details(self, uuid: str) -> bool:
-        return self.analysis_details.pop(uuid, None) is not None
+        with self.sync_lock:
+            return self.analysis_details.pop(uuid, None) is not None
 
     def analysis_details_exists(self, uuid: str) -> bool:
-        return uuid in self.analysis_details
+        with self.sync_lock:
+            return uuid in self.analysis_details
 
     def reset(self):
         self.root_analysis = {}
